chess_ai_mcts/requirements.txt
python-chess==1.10.0
numpy==1.24.3
matplotlib==3.7.2
pandas==2.0.3
